Refactor TTAlchemist class and external C API
