+++ tools/tt-alchemist/include/tt-alchemist/README.md
@@ -1,0 +1,1 @@
+This directory contains **external facing** headers for the tt-alchemist library.
